+++ src/main/java/page/objects/HomeFaqPage.java
@@ -2,6 +2,9 @@ package page.objects;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
 
 public class HomeFaqPage {
     private final WebDriver driver;
@@ -30,6 +33,10 @@ public class HomeFaqPage {
                By.xpath("//*[@id='accordion__panel-7']/p")
     };
 
+    private final By orderButtonDown = By.xpath(".//div[starts-with(@class, 'Order_Buttons')]/button[text()='Заказать'] ");
+
+    private final By orderButtonUp = By.xpath(".//div[starts-with(@class, 'Header_Nav')]/button[text()='Заказать']");
+
     public HomeFaqPage(WebDriver driver) {
         this.driver = driver;
     }
@@ -38,9 +45,28 @@ public void clickOnQuestion(int heading) {
     driver.findElement(headingLocators[heading]).click();
 }
 
+public void cookieButtonClick (WebDriverWait wait)
+{
+    var cookieButton = wait.until(ExpectedConditions.elementToBeClickable(By.id("rcc-confirm-button")));
+    cookieButton.click();
+}
+
 // Метод для получения ответа
     public String getAnAnswer(int panel) {
         return driver.findElement(panelLocators[panel]).getText();
     }
+
+    public OrderForm clickOrderButtonUp() {
+        driver.findElement(orderButtonUp).click();
+        return new OrderForm(driver);
+    }
+
+    // Метод для клика на нижнюю кнопку "Заказать"
+    public OrderForm clickOrderButtonDown() {
+        driver.findElement(orderButtonDown).click();
+        return new OrderForm(driver);
+    }
        }
+
+
